+++ modules/indicia_setup/db/version_9_6_0/202410270900_fields_for_import_reverser.sql.sql
@@ -1,0 +1,9 @@
+ALTER TABLE samples 
+ADD COLUMN import_guid 
+CHARACTER VARYING;
+COMMENT ON COLUMN samples.import_guid IS 'Globally unique identifier of the import batch.';
+
+ALTER TABLE imports
+ADD COLUMN reversible
+BOOLEAN default FALSE;
+COMMENT ON COLUMN samples.import_guid IS 'FALSE if reverse is not permitted (such as if import has already been reversed).';