+++ modules/summary_builder/config/summary_builder.php
@@ -118,7 +118,7 @@ $config['get_YearTaxonLocationUser_query'] = "
   SELECT 1 AS count, p.id AS sample_id, p.date_start, 't' as present
 	FROM occurrences o
 	JOIN samples s ON s.id = o.sample_id AND s.deleted = 'f'
-	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.created_by_id = #user_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.created_by_id = #user_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31' AND p.training = 'f'
 	WHERE o.taxa_taxon_list_id = #taxon_id#
 		AND o.deleted = 'f' AND o.zero_abundance = 'f'
   UNION ALL
@@ -129,6 +129,7 @@ $config['get_YearTaxonLocationUser_query'] = "
 		AND p.created_by_id = #user_id#
 		AND p.deleted = 'f'
 		AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+        AND p.training = 'f'
   ";
 
 $config['get_YearTaxonLocationUser_Attr_query'] = "
@@ -136,7 +137,7 @@ $config['get_YearTaxonLocationUser_Attr_query'] = "
 	FROM occurrences o
 	JOIN occurrence_attribute_values oav ON oav.occurrence_id = o.id AND oav.deleted = 'f' AND oav.occurrence_attribute_id = #attr_id#
 	JOIN samples s ON s.id = o.sample_id AND s.deleted = 'f'
-	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.created_by_id = #user_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.created_by_id = #user_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31' AND p.training = 'f'
 	WHERE o.taxa_taxon_list_id = #taxon_id#
 		AND o.deleted = 'f' AND o.zero_abundance = 'f'
   UNION ALL
@@ -147,13 +148,14 @@ $config['get_YearTaxonLocationUser_Attr_query'] = "
 		AND p.created_by_id = #user_id#
 		AND p.deleted = 'f'
 		AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+        AND p.training = 'f'
   ";
 
 $config['get_YearTaxonLocation_query'] = "
   SELECT 1 AS count, p.id AS sample_id, p.date_start, 't' as present
 	FROM occurrences o
 	JOIN samples s ON s.id = o.sample_id AND s.deleted = 'f'
-	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31' AND p.training = 'f'
 	WHERE o.taxa_taxon_list_id = #taxon_id#
 		AND o.deleted = 'f' AND o.zero_abundance = 'f'
   UNION ALL
@@ -163,6 +165,7 @@ $config['get_YearTaxonLocation_query'] = "
 		AND p.location_id = #location_id#
 		AND p.deleted = 'f'
 		AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+        AND p.training = 'f'
   ";
 
 $config['get_YearTaxonLocation_Attr_query'] = "
@@ -170,7 +173,7 @@ $config['get_YearTaxonLocation_Attr_query'] = "
 	FROM occurrences o
 	JOIN occurrence_attribute_values oav ON oav.occurrence_id = o.id AND oav.deleted = 'f' AND oav.occurrence_attribute_id = #attr_id#
 	JOIN samples s ON s.id = o.sample_id AND s.deleted = 'f'
-	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+	JOIN samples p ON s.parent_id = p.id AND p.survey_id = #survey_id# AND p.location_id = #location_id# AND p.deleted = 'f' AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31' AND p.training = 'f'
 	WHERE o.taxa_taxon_list_id = #taxon_id#
 		AND o.deleted = 'f' AND o.zero_abundance = 'f'
   UNION ALL
@@ -180,6 +183,7 @@ $config['get_YearTaxonLocation_Attr_query'] = "
 		AND p.location_id = #location_id#
 		AND p.deleted = 'f'
 		AND p.date_end>='#year#-01-01' AND p.date_start<='#year#-12-31'
+        AND p.training = 'f'
   ";
 
 $config['get_YearTaxonUser_query'] = "